+++ lib/BussinessLayer/Controllers/ware_controller.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
+import 'package:matjary/BussinessLayer/Controllers/connectivity_controller.dart';
 import 'package:matjary/BussinessLayer/Controllers/wares_controller.dart';
 import '../../DataAccesslayer/Models/ware.dart';
 import '../../DataAccesslayer/Repositories/ware_repo.dart';
@@ -12,6 +13,7 @@ class WareController extends GetxController {
   WaresController waresController = Get.find<WaresController>();
   var loading = false.obs;
   var savingState = false;
+  final connectivityController = Get.find<ConnectivityController>();
 
   void setWareName(name) {
     if (name.isNotEmpty) {
@@ -37,16 +39,20 @@ class WareController extends GetxController {
     savingState = false;
     String wareName = getWareName();
     if (wareName.isNotEmpty) {
-      loading.value = true;
-      Ware? ware = await wareRepo.craeteWare(wareName);
-      loading.value = false;
-      if (ware != null) {
-        waresController.getWares();
-        savingState = true;
-        setWareDetails(null);
-        SnackBars.showSuccess('تمت إضافة مستودع جديد');
+      if (connectivityController.isConnected) {
+        loading.value = true;
+        Ware? ware = await wareRepo.craeteWare(wareName);
+        loading.value = false;
+        if (ware != null) {
+          waresController.getWares();
+          savingState = true;
+          setWareDetails(null);
+          SnackBars.showSuccess('تمت إضافة مستودع جديد');
+        } else {
+          SnackBars.showError('حدث خطأ أثناء الإضافة');
+        }
       } else {
-        SnackBars.showError('حدث خطأ أثناء الإضافة');
+        SnackBars.showError('لا يوجد اتصال بالانترنت');
       }
     } else {
       SnackBars.showWarning('يرجى تعبئة الحقول المطلوبة');
@@ -56,15 +62,19 @@ class WareController extends GetxController {
   Future<void> updateWare(int id) async {
     String wareName = getWareName();
     if (wareName.isNotEmpty) {
-      setWareDetails(null);
-      loading.value = true;
-      var account = await wareRepo.updateWare(id, wareName);
-      loading.value = false;
-      if (account != null) {
-        waresController.getWares();
-        SnackBars.showSuccess('تم التعديل بنجاح');
+      if (connectivityController.isConnected) {
+        setWareDetails(null);
+        loading.value = true;
+        var account = await wareRepo.updateWare(id, wareName);
+        loading.value = false;
+        if (account != null) {
+          waresController.getWares();
+          SnackBars.showSuccess('تم التعديل بنجاح');
+        } else {
+          SnackBars.showError('فشل التعديل');
+        }
       } else {
-        SnackBars.showError('فشل التعديل');
+        SnackBars.showError('لا يوجد اتصال بالانترنت');
       }
     } else {
       SnackBars.showWarning('يرجى تعبئة الحقول المطلوبة');
@@ -72,14 +82,18 @@ class WareController extends GetxController {
   }
 
   Future<void> deleteWare(id) async {
-    loading.value = true;
-    var ware = await wareRepo.deleteWare(id);
-    loading.value = false;
-    if (ware != null) {
-      waresController.getWares();
-      SnackBars.showSuccess('تم الحذف بنجاح');
+    if (connectivityController.isConnected) {
+      loading.value = true;
+      var ware = await wareRepo.deleteWare(id);
+      loading.value = false;
+      if (ware != null) {
+        waresController.getWares();
+        SnackBars.showSuccess('تم الحذف بنجاح');
+      } else {
+        SnackBars.showError('فشل الحذف');
+      }
     } else {
-      SnackBars.showError('فشل الحذف');
+      SnackBars.showError('لا يوجد اتصال بالانترنت');
     }
   }
 }
